Extract source metadata mutation executor
